+++ src/dev/build.ts
@@ -14,7 +14,7 @@ import {
   join,
 } from "../deps.ts";
 import { fileExists } from "../server/utils.ts";
-import type { GetRouterOptions } from "../server/router.ts";
+import type { GetRouterOptions, Handlers } from "../server/router.ts";
 import type { AppTemplateInterface } from "../server/ssr.ts";
 
 const TEST_FILE_PATTERN = /[._]test\.(?:[tj]sx?|[mc][tj]s)$/;
@@ -24,7 +24,7 @@ export type BuildRoute = {
   path: string;
   relativeFilePath: string;
   absoluteFilePath: string;
-  routeModule?: { default: unknown };
+  routeModule?: { default: unknown; handler?: Handlers };
   tagName: string;
   jsAssetContent: esbuild.OutputFile | undefined;
   jsAssetPath: string | undefined;
@@ -258,16 +258,22 @@ export async function getRoutes({
 }
 
 export async function getAppTemplate({ loadFs }: GetRouterOptions) {
-  const hasAppJs = await fileExists("./routes/_app.js").catch(() => false);
-  const hasAppTs = await fileExists("./routes/_app.ts").catch(() => false);
-  if (hasAppJs && hasAppTs) {
+  const [checkTs, checkJs] = await Promise.allSettled([
+    fileExists("./routes/_app.ts"),
+    fileExists("./routes/_app.js"),
+  ]);
+
+  const hasAppTs = checkTs.status === "fulfilled" && checkTs.value === true;
+  const hasAppJs = checkJs.status === "fulfilled" && checkJs.value === true;
+
+  if (hasAppTs && hasAppJs) {
     throw new Error(
-      "You have two app templates defined: _app.js and _app.ts. Use only one."
+      "You have two app templates defined: _app.ts and _app.js. Use only one."
     );
   }
 
-  if (!hasAppJs && !hasAppTs) {
-    throw new Error("You don't an app template defined: _app.js and _app.ts.");
+  if (!hasAppTs && !hasAppJs) {
+    throw new Error("You don't an app template defined: _app.ts or _app.js.");
   }
 
   if (hasAppTs) {