+++ src/server/ssr.ts
@@ -1,4 +1,5 @@
-import { html, unsafeHTML } from "../deps.ts";
+import { html, unsafeHTML, render, collectResult } from "../deps.ts";
+import type { RouterContext } from "../deps.ts";
 // @ts-ignore lit is a npm package and Deno doesn't resolve the exported members
 import type { LitElement, TemplateResult } from "lit";
 // @ts-ignore lit is a npm package and Deno doesn't resolve the exported members
@@ -6,6 +7,8 @@ import type { DirectiveResult } from "lit/directives/unsafe-html.js";
 // @ts-ignore lit is a npm package and Deno doesn't resolve the exported members
 import type { UnsafeHTMLDirective } from "lit/directives/unsafe-html.js";
 import type { BuildRoute } from "../dev/build.ts";
+import { LimetteElementRenderer } from "./rendering/limette-element-renderer.ts";
+import type { ComponentContext } from "./router.ts";
 
 import "../runtime/is-land.ts"; // should use limette?
 
@@ -14,7 +17,7 @@ type Params = {
 };
 
 export type AppTemplateOptions = {
-  css: string;
+  css: DirectiveResult<UnsafeHTMLDirective> | string;
   js: string[] | TemplateResult[] | DirectiveResult<UnsafeHTMLDirective>[];
   component: DirectiveResult<UnsafeHTMLDirective>;
 };
@@ -28,6 +31,16 @@ export declare class AppTemplateInterface {
   prototype: {
     render(app: AppTemplateOptions): TemplateResult;
   };
+}
+
+function registerComponent(module: CustomElementConstructor, tagName: string) {
+  if (!customElements.get(`lmt-route-${tagName}`)) {
+    customElements.define(`lmt-route-${tagName}`, module);
+  } else {
+    // check if it is the same class
+  }
+
+  return `<lmt-route-${tagName}></lmt-route-${tagName}>`;
 }
 
 const ComponentCtxMixin = (
@@ -69,29 +82,27 @@ const ComponentCtxMixin = (
 export function bootstrapContent(
   AppTemplate: AppTemplateInterface,
   route: BuildRoute,
-  ctx: {
-    params: Params;
-  }
+  componentContext: ComponentContext
 ) {
   const componentModule = route.routeModule;
 
   const componentClass = ComponentCtxMixin(
     componentModule?.default as typeof LitElement,
-    {
-      params: ctx.params,
-    }
+    componentContext
   );
   const component = registerComponent(
     componentClass as unknown as CustomElementConstructor,
     route.tagName
   );
 
-  const ctxStr = `<script type="text/json" id="_lmt_ctx">${JSON.stringify({
-    params: ctx.params,
-  })}</script>`;
+  const ctxStr = `<script type="text/json" id="_lmt_ctx">${JSON.stringify(
+    componentContext
+  )}</script>`;
 
   const appTemplateOptions: AppTemplateOptions = {
-    css: route.cssAssetPath ? route.cssAssetPath : ``,
+    css: route.cssAssetPath
+      ? unsafeHTML(`<link rel="stylesheet" href="${route.cssAssetPath}" />`)
+      : ``,
     js: [
       route.jsAssetPath ? unsafeHTML(ctxStr) : ``,
       route.jsAssetPath
@@ -104,12 +115,24 @@ export function bootstrapContent(
   return AppTemplate.prototype.render(appTemplateOptions);
 }
 
-function registerComponent(module: CustomElementConstructor, tagName: string) {
-  if (!customElements.get(`lmt-route-${tagName}`)) {
-    customElements.define(`lmt-route-${tagName}`, module);
-  } else {
-    // check if it is the same class
-  }
+export async function renderContent(
+  AppTemplate: AppTemplateInterface,
+  route: BuildRoute,
+  routerContext: RouterContext<typeof route.path>,
+  data?: ComponentContext["data"]
+) {
+  const componentContext = { params: routerContext.params, data };
 
-  return `<lmt-route-${tagName}></lmt-route-${tagName}>`;
+  const result = render(
+    await bootstrapContent(
+      AppTemplate as AppTemplateInterface,
+      route,
+      componentContext
+    ),
+    {
+      elementRenderers: [LimetteElementRenderer(route)],
+    }
+  );
+  const content = await collectResult(result);
+  return content;
 }